requirements.txt

streamlit
numpy
tensorflow
Pillow
requests
joblib
